refactor(schema): share client IP derivation
Moves getIpAddress and maskIpAddress into @c15t/schema alongside the geo
header helpers — same category of thing, request metadata recovered from proxy
headers — and has @c15t/backend delegate.

Shared rather than duplicated because the derived IP is written to every
consent record. That makes it a compliance field, not a diagnostic one: two
backends deriving different values for the same visitor during the parallel
phase would put inconsistent data in the audit trail, and unlike a latency
difference that cannot be corrected afterwards. The IPv6 expand/compress
masking is also fiddly enough that a second implementation would drift.

Behaviour preserved: the shipping package's own 24 process-ip tests pass
against the delegated version, and a parity test covers 36 combinations of
header, forwarding chain and configuration.

Defaults are unchanged and worth restating: masking on unless explicitly
disabled, tracking:false records no IP at all rather than a placeholder, and
header precedence puts x-client-ip ahead of x-forwarded-for.

backend-next now accepts ipAddress config and 80 tests pass; 566 backend and
116 schema tests unchanged.
